fix inquirePowerDetail

+++ oas-server/oas/src/main/java/com/cascv/oas/server/energy/vo/EnergyPowerChangeDetail.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 public class EnergyPowerChangeDetail {
 	@Getter @Setter private String uuid;
 	@Getter @Setter private String userUuid;
-	@Getter @Setter private Integer value;
+	@Getter @Setter private BigDecimal value;
 	@Getter @Setter private BigDecimal powerChange;
 	@Getter @Setter private String category;
 	@Getter @Setter private String source;
